Coerce to a data frame.

Functions to coerce an intensity object to a data frame.

# S3 method for intensity
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = FALSE
)

Arguments

x

An intensity object.

row.names

NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional

logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

...

additional arguments to be passed to or from methods.

stringsAsFactors

logical: should the character vector be converted to a factor?

Value

A data frame.

Examples

my_data <- incidence(tomato_tswv$field_1929)
head(as.data.frame(my_data))
#>   x y t i n
#> 1 1 1 1 0 1
#> 2 1 2 1 1 1
#> 3 1 3 1 0 1
#> 4 1 4 1 1 1
#> 5 1 5 1 0 1
#> 6 1 6 1 0 1