C(alpha) test.

The C(alpha) test is a test of the binomial distribution against the alternative of the beta-binomial distribution.

calpha.test(x, ...)

# S3 method for fisher
calpha.test(x, ...)

Arguments

x
The output of the agg_index function with method = "fisher" as parameter.

...
Not yet implemented.

Details

It is based on calculation of a test statistic, z, that has an asymptotic standard normal distribution under the null hypothesis. It is one-sided (in the way that the alternative is aggregation, not just "non-randomness"), thus with a confidence level of 95 1.64. When all the sampling units contain the same total number of individuals, n, the test statistic is calculated from:

z = (n(N - 1)I - Nn)/(2Nn(n - 1))^(1/2)

where N is the number of sampling units, and I, Fisher's index of aggregation for incidence data.

References

Neyman J. 1959. Optimal asymptotic tests of composite statistical hypotheses. In: Probability and Statistics, 213-234. Wiley, New York.

Tarone RE. 1979. Testing the goodness of fit of the binomial distribution. Biometrika, 66(3): 585-590.

See also

Examples

# For incidence data:
my_incidence <- incidence(tobacco_viruses)
my_fisher <- agg_index(my_incidence, method = "fisher")
calpha.test(my_fisher)

#> 
#> 	C(alpha) test
#> 
#> data:  my_fisher
#> z = 13.036, p-value < 2.2e-16
#>