Count of codling moth larvae.

Codling moth diapausing larvae were collected in an apple orchard in south-eastern France. Larvae were caught on strip traps wrapped around tree trunks in July 2008 and collected the following October. 30 traps were used.

codling_moths

Format

A data frame with 30 rows and 3 variables:

[, 1:2]	x,y	Metric spatial coordonates.

[, 3]	i	Counts of codling moth larvae.

Source

Lavigne C, Ricci B, Franck P, Senoussi R. 2010. Spatial analyses of ecological count data: A density map comparison approach. Basic and Applied Ecology 11: 734-42. doi:10.1016/j.baae.2010.08.011