Incidence of three viruses in an Australian hop garden.

Three viruses, i.e. Hop latent virus (HpLV), Hop mosaic virus (HpMV), and Apple mosaic virus (ApMV), were monitored in an Australian hop garden for two consecutive years (1996 and 1997). The hop garden was established in 1989 with the variety Victoria in a commercial hop farm at Bushy Park, Tasmania, Australia. It consisted of 25 rows containing 51 plants each, so that there were 1275 hop plants in total. There were 2.1 m between rows, and 1.8 m between plants within rows.

hop_viruses

Format

There are three data frames, one for each virus (HpLV, HpMV and ApMV). Each data frame consists of 2550 rows and 7 variables:

[, 1:2]	x,y	Grid spatial coordonates.

[, 3:4]	xm,ym	Metric spatial coordinates.

[, 5]	t	Year of disease assessments.

[, 6]	i	Incidence. 0: Healthy, 1: Diseased.

[, 7]	n	Sampling unit size. n = 1 means that the sampling unit size is the plant.

Source

Pethybridge SJ, Madden LV. 2003. Analysis of spatiotemporal dynamics of virus spread in an Australian hop garden by stochastic modeling. Plant Disease 87:56-62. doi:10.1094/PDIS.2003.87.1.56