Retrieve vector or array indices

ind2sub is a synonym for arrayInd in base package.

Usage

ind2sub(ind, .dim, .dimnames = NULL, useNames = FALSE)

sub2ind(sub, .dim)

Arguments

ind

Vector indices.

.dim

dim(.) integer vector

.dimnames

optional list of character dimnames(.). If useNames is true, to be used for constructing dimnames for arrayInd() (and hence, which(*, arr.ind=TRUE)). If names(.dimnames) is not empty, these are used as column names. .dimnames[[1]] is used as row names.

useNames

logical indicating if the value of arrayInd() should have (non-null) dimnames at all.

sub

Array/matrix indices.

Details

ind2sub is just an alias for arrayInd

Examples

set.seed(12345)
mat <- matrix(round(runif(6, min = 0, max = 10)), nrow = 2, ncol = 3)
ind2sub(4, dim(mat))
#>      [,1] [,2]
#> [1,]    2    2
sub2ind(c(2, 2), dim(mat))
#> [1] 4
subs <- as.matrix(expand.grid(1:2,2:3))
sub2ind(subs, dim(mat))
#> [1] 3 4 5 6