Extract log-likelihood

This function returns the maximal log-likelihood estimated with smle, if f returned log-likelihood value.

# S3 method for smle
logLik(object, ...)

Arguments

object

any object from which a log-likelihood value, or a contribution to a log-likelihood value, can be extracted.

...

some methods for this generic function require additional arguments.

Value

Returns an object of class logLik. This is a number with at least one attribute, "df" (degrees of freedom), giving the number of (estimated) parameters in the model.

There is a simple print method for "logLik" objects.

There may be other attributes depending on the method used: see the appropriate documentation. One that is used by several methods is

"nobs", the number of observations used in estimation (after the restrictions if REML = TRUE).