Construct data mappings.

Data mappings describe how variables in the data are mapped to standard names used throughout epiphy.

mapping(...)

mapping_(x)

remap(data, mapping, keep_only_std = TRUE)

Arguments

...

One or more unquoted expressions separated by commas.

x

Vector of one or more character strings.

data

An intensity object.

mapping

A mapping object.

keep_only_std

Keep only standard variables.

Value

A list of mapped names.

Details

Standard names are x, y and z for the three spatial dimensions, and t for the time. r corresponds to the records of (disease) intensity, and n, the number of individuals in a sampling unit (if applicable).

mapping() works with expressions, and mapping_(), with a vector of characters.

See also

Examples

mapping(x = col1, y = col2)
#> * x -> col1
#> * y -> col2
mapping_(c("x = col1", "y = col2"))
#> * x -> col1
#> * y -> col2