Incidence of tomato spotted wilt virus (TSWV) disease in field trials.

Intensively mapped TSWV incidence data reported by Cochran (1936) and Bald (1937). The disease assessments were performed in field trials at the Waite Institute (Australia) in 1928 and 1929. TSWV is a virus disease spread by thrips.

Usage

tomato_tswv

Format

There are two data frames:

field_1928: A data frame with 11088 rows and 8 variables:

[, 1]	plot	Plot id.

[, 2]	variety	Variety name.

[, 3]	irrigation	Irrigation system.

[, 4:5]	x,y	Grid spatial coordinates.

[, 6]	t	Date of disease assessments. 1: 6 Nov, 2: 14 Nov, 3: 21 Nov, 4: 28-29 Nov, 5: 5 Dec, 6: 12 Dec 1928.

[, 7]	i	Disease incidence. 0: Healthy, 1: Diseased.

[, 8]	n	Sampling unit size. n = 1 means that the sampling unit size is the plant.

field_1929: A data frame with 4320 rows and 5 variables:

[, 1:2]	x,y	Grid spatial coordinates.

[, 3]	t	Date of disease assessments. 1: 18 Dec, 2: 31 Dec 1929, 3: 22 Jan 1930.

[, 4]	i	Disease incidence. 0: Healthy, 1: Diseased.

[, 5]	n	Sampling unit size. n = 1 means that the sampling unit size is the plant.

Source

Cochran WG. 1936. The statistical analysis of field counts of diseased plants. Supplement to the Journal of the Royal Statistical Society 3, 49–67. doi:10.2307/2983677

Bald JG. 1937. Investigations on "spotted wilt" of tomatoes. III. Infection in field plots. Bulletin 106. Melbourne, Australia: Council for Scientific and Industrial Research.

Details

The data set field_1928, reported by Bald (1937), was a set of four plots. Each plot consisted of 14 rows containing 33 plants each, so that there were 462 plants in each plot. The tomato variety Early Dwarf Red was used in two plots, and the variety Burwood Prize in the other two. The tomatoes were planted out on 15th October 1928. The two plots dedicated to a given variety experienced different irrigation practices, using either overhead sprays or trenches. Otherwise, all were treated alike. Weekly records of TSWV incidence were performed from 6th November to 12th December.

The data set field_1929, reported by Cochran (1936), was a field of 24 rows containing 60 plants each, so that there were 1440 plants. The tomatoes were planted out in 26th November 1929. TSWV incidence records made on 18th December 1929, 31st December 1929 and 22nd January 1930 are reported in this data set.