Calculate Variance-Covariance Matrix for a Fitted Model Object

Returns the variance-covariance matrix of the main parameters of a fitted model object. The “main” parameters of model correspond to those returned by coef, and typically do not contain a nuisance scale parameter (sigma).

# S3 method for smle
vcov(object, ...)

Arguments

object

a fitted model object, typically. Sometimes also a summary() object of such a fitted model.

...

additional arguments for method functions. For the glm method this can be used to pass a dispersion parameter.

Value

A matrix of the estimated covariances between the parameter estimates in the linear or non-linear predictor of the model. This should have row and column names corresponding to the parameter names given by the

coef method.

When some coefficients of the (linear) model are undetermined and hence NA because of linearly dependent terms (or an

“over specified” model), also called

“aliased”, see alias, then since R version 3.5.0,

vcov() (iff complete = TRUE, i.e., by default for

lm etc, but not for aov) contains corresponding rows and columns of NAs, wherever coef() has always contained such NAs.