Calculate Variance-Covariance Matrix for a Fitted Model Object

Returns the variance-covariance matrix of the main parameters of a fitted model object.

# S3 method for smle
vcov(object, ...)

Arguments

object
a fitted model object, typically. Sometimes also a summary() object of such a fitted model.

...
additional arguments for method functions. For the glm method this can be used to pass a dispersion parameter.

Value

A matrix of the estimated covariances between the parameter estimates in the linear or non-linear predictor of the model. This should have row and column names corresponding to the parameter names given by the coef method.